Move `type` to `ArrayMerge`
