+++ src/Expression/Function/ArrayMerge.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Expression\Function;
 
+use Yiisoft\Db\Schema\Column\ColumnInterface;
+
 /**
  * Represents an SQL expression that returns the merged array from a list of operands.
  *
@@ -21,4 +23,22 @@ namespace Yiisoft\Db\Expression\Function;
  */
 final class ArrayMerge extends MultiOperandFunction
 {
+    private string|ColumnInterface $type = '';
+
+    /**
+     * Returns the type of the operands. Empty string if not set.
+     */
+    public function getType(): string|ColumnInterface
+    {
+        return $this->type;
+    }
+
+    /**
+     * Sets the type of the operands.
+     */
+    public function type(string|ColumnInterface $type): self
+    {
+        $this->type = $type;
+        return $this;
+    }
 }
